% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_estaciones.R
\docType{data}
\name{datos_estaciones}
\alias{datos_estaciones}
\title{Titulo}
\format{
\describe{
\item{}{}
\item{}{}
}
}
\source{
datos que diponibiliza el Institulo Nacional de Tecnología Agropecuaria INTA
a traves del Sistema de Información y Gestión Agropecuaria SIGA.
}
\usage{
datos_estaciones
}
\description{
Describir
}
\keyword{datasets}
