% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Gráfico del promedio mensual de la temperatura}
\usage{
grafico_temperatura_mensual(
  dataframe,
  id,
  color = "magenta",
  titulo = "Temperatura"
)
}
\arguments{
\item{dataframe}{nombre del dataframe a utilizar}

\item{id}{clave unica de identificacion de la estacion agroclimatica}

\item{color}{(opcional) color que desea utilizar en el gráfico, expresado en
rgb, hsvy, hcl, o el nombre del color en ingles}

\item{título}{(opcional) el nombre asignado para el gráfico escrito entre comillas}
}
\value{
grafico de puntos indicando el promedio de la temperatura a lo largo del año
}
\description{
La función \code{grafico_temperatura_mensual()} toma como argumento un dataframe,
calcula el promedio de la temperatura de cada mes y lo grafica
}
\examples{
grafico_temperatura_mensual(datos_estaciones,color = "magenta", titulo = "Temperatura promedio mensual")
grafico_temperatura_mensual(datos_estaciones, título = "Temperatura de todos los meses")
grafico_temperatura_mensual(datos_estaciones,color = "blue")
grafico_temperatura_mensual(datos_estaciones)

}
