% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadatos.R
\docType{data}
\name{metadatos}
\alias{metadatos}
\title{Metadatos}
\format{
\describe{
\item{id}{Clave única de identificación de la estacion agroclimática}
\item{nombre}{Nombre de la estación agroclimática}
\item{tipo}{Tipo de estación}
\item{localidad}{Nombre de la localidad en la que esta ubicada la estación}
\item{provincia}{Nombre de la provincia en la que se encuentra la estación}
\item{lat}{Latitud de la ubicacion de la estación}
\item{lon}{Longitud de la ubicacion de la estación}
\item{altura}{Altura de la estación sobre el nivel del mar}
\item{ubicacion}{Dirección exacta de la ubicacion de la estación}
\item{desde}{Fecha en la que se comenzó a recopilar datos en la estación}
\item{hasta}{Última fecha en la que recopilaron datos en la estación}
}
}
\source{
Datos que diponibiliza el Institulo Nacional de Tecnología Agropecuaria INTA
a traves del Sistema de Información y Gestión Agropecuaria SIGA.
}
\usage{
metadatos
}
\description{
Datos sobre estaciones agroclimáticas INTA (Instituto Nacional de Tecnología Agropecuaria )
}
\keyword{datasets}
