% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NH0098.R
\docType{data}
\name{NH0098}
\alias{NH0098}
\title{NH0098}
\format{
\describe{
\item{id}{Clave unica de identificacion de la estacion }
\item{fecha}{Fecha de la observacion}
\item{temperatura_abrigo_150cm}{Temperatura medida a 150 cm del suelo, en grados Celsius.}
\item{temperatura_abrigo_150cm_maxima}{Temperatura maxima medida a 150 cm del suelo, en grados Celsius.}
\item{temperatura_abrigo_150cm_minima}{Temperatura minima medida a 150 cm del suelo, en grados Celsius.}
\item{temperatura_intemperie_5cm_minima}{Temperatura minima medida al aire libre a 5 cm del suelo, en grados Celsius.}
\item{temperatura_intemperie_50cm_minima}{Temperatura minima medida al aire libre a 50 cm del suelo, en grados Celsius.}
\item{temperatura_suelo_5cm_media}{Temperatura media del suelo a 5cm de profundidad, en grados Celsius.}
\item{temperatura_suelo_10cm_media}{Temperatura media del suelo a 10 cm de profundidad, en grados Celsius.}
\item{temperatura_inte_5cm}{Temperatura medida al aire libre a 5 cm del suelo, en grados Celsius.}
\item{temperatura_intemperie_150cm_minima}{Temperatura minima medida al aire libre a 150 cm del suelo, en grados Celsius.}
\item{humedad_suelo}{Humedad medida en el suelo.}
\item{precipitacion_pluviometrica}{Cantidad de precipitacion acumulada, en milimetros.}
\item{granizo}{1: Si, 0:No.}
\item{nieve}{1: Si, 0:No.}
\item{heliofania_efectiva}{Horas de sol efectivo al dia.}
\item{heliofania_relativa}{Porcentaje de sol efectivo respecto a las horas de luz disponibles.}
\item{tension_vapor_media}{Tension de vapor promedio, en hectopascales.}
\item{humedad_media}{Humedad relativa promedio.}
\item{humedad_media_8_14_20}{Humedad relativa medida a las 8, 14 y 20 horas.}
\item{rocio_medio}{Punto de rocio medio}
\item{duracion_follaje_mojado}{Horas de follaje mojado por dia.}
\item{velocidad_viento_200cm_media}{Velocidad media del viento a 200 cm del suelo, en km por hora.}
\item{direccion_viento_200cm}{Direccion del viento a 200 cm del suelo, en grados. }
\item{velocidad_viento_1000cm_media}{Velocidad media del viento a 1000 cm del suelo, en km por hora.}
\item{direccion_viento_1000cm}{Direccion del viento a 1000 cm del suelo, en grados.}
\item{velocidad_viento_maxima}{Velocidad maxima del viento regristrada, en km por hora.}
\item{presion_media}{Presion atmosferica media, en hectopascales.}
\item{radiacion_global}{Cantidad de radiacion solar global.}
\item{radiacion_neta}{Cantidad de radiacion solar neta. }
\item{evaporacion_tanque}{Evaporacion de agua diaria medida en un tanque, en milimetros,}
\item{evapotranspiracion_potencial}{Evapotranspiracion potencial diaria, en milimetros .}
\item{profunidad_napa}{Profunidad de la napa freatica, en metros.}
\item{horas_frio}{Cantidad de horas de frio acumuladas.}
\item{unidad_frio}{Unidades de frio acumuladas.}

}
}
\source{
Datos que diponibiliza el Institulo Nacional de Tecnología Agropecuaria INTA
a traves del Sistema de Información y Gestión Agropecuaria SIGA.
}
\usage{
NH0098
}
\description{
Conjunto de datos recopilados en la estacion NH0098 del INTA
}
\keyword{datasets}
