% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Tabla de resumen}
\usage{
tabla_resumen_temperatura(data, id)
}
\arguments{
\item{data}{nombre del dataframe a utilizar}

\item{id}{clave unica de identificacion de la estacion agroclimatica}
}
\value{
tabla que contiene un resumen de los datos de la temperatura
}
\description{
La funcion \code{tabla_resumen_temperatura()} toma como argumento un dataframe y
resume los datos de la temperatura
}
\examples{
tabla_resumen_temperatura(datos_estaciones, "NH0910")
tabla_resumen_tempertura(datos_estaciones)


}
