% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Gráfico del promedio mensual de la temperatura}
\usage{
grafico_temperatura_mensual(
  dataframe,
  color = "magenta",
  titulo = "Temperatura"
)
}
\arguments{
\item{dataframe}{nombre del dataframe a utilizar}

\item{color}{(opcional) color que desea utilizar en el gráfico, expresado en
rgb, hsvy, hcl, o el nombre del color en ingles}

\item{título}{(opcional) el nombre asignado para el gráfico escrito entre comillas}
}
\value{
gráfico de puntos indicando el promedio de la temperatura a lo largo del año
}
\description{
La función \code{grafico_temperatura_mensual()} toma como argumento un dataframe,
calcula el promedio de la temperatura de cada mes y lo grafica
}
\examples{
grafico_temperatura_mensual(NH0910,color = "magenta", título = "Temperatura promedio mensual")
grafico_temperatura_mensual(NH0098, título = "Temperatura de todos los meses")
grafico_temperatura_mensual(NH0046,color = "blue")
grafico_temperatura_mensual(NH0472)

}
